+++ objectbox-java/src/main/java/io/objectbox/BoxStore.java
@@ -82,7 +82,7 @@ public class BoxStore implements Closeable {
      * ReLinker uses this as a suffix for the extracted shared library file. If different, it will update it. Should be
      * unique to avoid conflicts.
      */
-    public static final String JNI_VERSION = "5.0.0-2025-09-27";
+    public static final String JNI_VERSION = "5.1.0-2026-01-19";
 
     /** The ObjectBox database version this Java library is known to work with. */
     private static final String VERSION = "5.1.0-2026-01-19";