+++ CHANGELOG.md
@@ -6,8 +6,8 @@ For more insights into what changed in the ObjectBox C++ core, [check the Object
 
 ## 5.0.1 - 2025-09-30
 
-- Include runtime libraries for Android and JVM with database version `5.0.0-2025-09-27`.
-- Fixed a race condition with a closing store and still active transactions that kept the store from closing.
+- Update runtime libraries for Android and JVM to database version `5.0.0-2025-09-27`.
+- Fix a race condition with a closing store and still active transactions that kept the store from closing.
   For Android this may fix some rare ANR issues.
 
 ### Sync